Finishing MAG L1A processing and tests
