+++ imap_processing/mag/l1a/mag_l1a_data.py
@@ -2,6 +2,8 @@
 from dataclasses import dataclass
 from math import floor
 
+import numpy as np
+
 MAX_FINE_TIME = 65535  # maximum 16 bit unsigned int
 
 
@@ -59,11 +61,17 @@ class Vector:
 
     # TODO: This timestamp should be in J2000/datetime64
     timestamp: TimeTuple
-    vectors: tuple[int, int, int, int]
+    x: int
+    y: int
+    z: int
+    rng: int
 
     def __init__(self, vectors, time: TimeTuple):
-        self.vectors = vectors
         self.timestamp = time
+        self.x = vectors[0]
+        self.y = vectors[1]
+        self.z = vectors[2]
+        self.rng = vectors[3]
 
 
 @dataclass
@@ -121,3 +129,125 @@ class MagL1a:
             previous_time = new_vector.timestamp
 
             self.vectors[index] = new_vector
+
+    @staticmethod
+    def process_vector_data(
+        vector_data: np.ndarray, primary_count: int, secondary_count: int
+    ) -> (list[tuple], list[tuple]):
+        """
+        Given raw packet data, process into Vectors.
+
+        Vectors are grouped into primary sensor and secondary sensor, and returned as a
+        tuple (primary sensor vectors, secondary sensor vectors)
+
+        Written by MAG instrument team
+
+        Parameters
+        ----------
+        vector_data : np.ndarray
+            Raw vector data, in bytes. Contains both primary and secondary vector data
+            (first primary, then secondary)
+        primary_count : int
+            Count of the number of primary vectors
+        secondary_count : int
+            Count of the number of secondary vectors
+
+        Returns
+        -------
+        (primary, secondary)
+            Two arrays, each containing tuples of (x, y, z, sample_range) for each
+            vector sample.
+        """
+
+        # TODO: error handling
+        def to_signed16(n):
+            n = n & 0xFFFF
+            return n | (-(n & 0x8000))
+
+        pos = 0
+        primary_vectors = []
+        secondary_vectors = []
+
+        for i in range(primary_count + secondary_count):  # 0..63 say
+            x, y, z, rng = 0, 0, 0, 0
+            if i % 4 == 0:  # start at bit 0, take 8 bits + 8bits
+                # pos = 0, 25, 50...
+                x = (
+                    ((vector_data[pos + 0] & 0xFF) << 8)
+                    | ((vector_data[pos + 1] & 0xFF) << 0)
+                ) & 0xFFFF
+                y = (
+                    ((vector_data[pos + 2] & 0xFF) << 8)
+                    | ((vector_data[pos + 3] & 0xFF) << 0)
+                ) & 0xFFFF
+                z = (
+                    ((vector_data[pos + 4] & 0xFF) << 8)
+                    | ((vector_data[pos + 5] & 0xFF) << 0)
+                ) & 0xFFFF
+                rng = (vector_data[pos + 6] >> 6) & 0x3
+                pos += 6
+            elif i % 4 == 1:  # start at bit 2, take 6 bits, 8 bit, 2 bits per vector
+                # pos = 6, 31...
+                x = (
+                    ((vector_data[pos + 0] & 0x3F) << 10)
+                    | ((vector_data[pos + 1] & 0xFF) << 2)
+                    | ((vector_data[pos + 2] >> 6) & 0x03)
+                ) & 0xFFFF
+                y = (
+                    ((vector_data[pos + 2] & 0x3F) << 10)
+                    | ((vector_data[pos + 3] & 0xFF) << 2)
+                    | ((vector_data[pos + 4] >> 6) & 0x03)
+                ) & 0xFFFF
+                z = (
+                    ((vector_data[pos + 4] & 0x3F) << 10)
+                    | ((vector_data[pos + 5] & 0xFF) << 2)
+                    | ((vector_data[pos + 6] >> 6) & 0x03)
+                ) & 0xFFFF
+                rng = (vector_data[pos + 6] >> 4) & 0x3
+                pos += 6
+            elif i % 4 == 2:  # start at bit 4, take 4 bits, 8 bits, 4 bits per vector
+                # pos = 12, 37...
+                x = (
+                    ((vector_data[pos + 0] & 0x0F) << 12)
+                    | ((vector_data[pos + 1] & 0xFF) << 4)
+                    | ((vector_data[pos + 2] >> 4) & 0x0F)
+                ) & 0xFFFF
+                y = (
+                    ((vector_data[pos + 2] & 0x0F) << 12)
+                    | ((vector_data[pos + 3] & 0xFF) << 4)
+                    | ((vector_data[pos + 4] >> 4) & 0x0F)
+                ) & 0xFFFF
+                z = (
+                    ((vector_data[pos + 4] & 0x0F) << 12)
+                    | ((vector_data[pos + 5] & 0xFF) << 4)
+                    | ((vector_data[pos + 6] >> 4) & 0x0F)
+                ) & 0xFFFF
+                rng = (vector_data[pos + 6] >> 2) & 0x3
+                pos += 6
+            elif i % 4 == 3:  # start at bit 6, take 2 bits, 8 bits, 6 bits per vector
+                # pos = 18, 43...
+                x = (
+                    ((vector_data[pos + 0] & 0x03) << 14)
+                    | ((vector_data[pos + 1] & 0xFF) << 6)
+                    | ((vector_data[pos + 2] >> 2) & 0x3F)
+                ) & 0xFFFF
+                y = (
+                    ((vector_data[pos + 2] & 0x03) << 14)
+                    | ((vector_data[pos + 3] & 0xFF) << 6)
+                    | ((vector_data[pos + 4] >> 2) & 0x3F)
+                ) & 0xFFFF
+                z = (
+                    ((vector_data[pos + 4] & 0x03) << 14)
+                    | ((vector_data[pos + 5] & 0xFF) << 6)
+                    | ((vector_data[pos + 6] >> 2) & 0x3F)
+                ) & 0xFFFF
+                rng = (vector_data[pos + 6] >> 0) & 0x3
+                pos += 7
+
+            vector = (to_signed16(x), to_signed16(y), to_signed16(z), rng)
+            if i < primary_count:
+                primary_vectors.append(vector)
+            else:
+                secondary_vectors.append(vector)
+
+        return (primary_vectors, secondary_vectors)
